Refactoring docs (sidebar in layout)
